Another way of delegate has been added but not tested. This is close to typical implementation of mix-in using method missing but it can be extended for general purpose delegate including local, REST, Thrift etc. The difference with Delegatable is as follows:
a user of fw can use both delegatable and delegate on instance val
if only for delegatable, there may be a lot of number of config file.
but it is difficult to make general purpose delegate such as delegate
with REST or Thrift. Therefore, although it is not simple and beautiful,
for practical purpose, both way of
of delegate is used.