Author thought it could be better to introduce type hint for most values. Since it makes mataintenace better and less document for app developer is required. However, for meta-programming, it might be better no to use type hints.
If type hints is required, C# is better way? Actually, it seems there
are useful meta-programming technique in C# and .NET. However, there is
type hint for PHP though there is no generics. There is JIT. There is
phar. And meta-programming of PHP is very simple and useful. The
question for meta-programming and C# is, if ORM is made by C# is it
possible to implement ORM with type?  So, considering from many
aspect, author thinks PHP is better way.